requirements.txt
alpaca-py>=0.13.0
transformers>=4.35.0
torch>=2.0.0
scipy>=1.11.0
requests>=2.31.0
pyyaml>=6.0.1
pandas>=2.1.0
schedule>=1.2.0
python-dotenv>=1.0.0
colorlog>=6.7.0
flask>=3.0.0
python-dateutil>=2.8.2
